Re-implement InitialReset
The InitialReset provided by the BSC stdlib is not synthesizable. This module is
appealing for use with iCE40 devices though, so this diff re-implements a
variant of this module.

Simply instantiate in the top of a design and use the 'reset_by' argument when
instantiating subsequent modules for design reset on PoR on iCE40 devices to
work as expected.