Allow Bluespec Verilog targets to be concrete
Sometimes it is desirable to generate Verilog output from a Bluespec module
without taking it all the way through synthesis for a concrete target. This diff
lets you pass an environment to such a target, making it concrete and causing it
to show up in the latest/ symlinks.

Example:

bluespec_verilog('SomeModule_verilog',
    env = 'default',
    top = 'SomeModule.bsv',
    modules = [
        'mkMyTop',
    ],
    deps = [
        ':SomeModule',
    ])

Alternate BSC flags can be set by using a different 'env'.