Add Bluesim test runner
This diff adds a simple Bluesim test runner, making it easier to build and run
the equivalent of unit test suites. Examples:

```
./cobble bluesim_test cobalt//.*Test
./cobble bluesim_test //hdl:SchmittRegTests.*
```
The first command builds and runs all Bluesim scripts where the name ends in
Test, providing a way to run all tests in the project. The second command builds
and runs only the tests in the SchmittRegTests suite.