Replace 8B10B encoding/decoding with Chuck Benz implementation
This diff is a significant overhaul of the 8B10B encoder, replacing it with the
implementation from Chuck Benz (see comments in source file). In addition both
the serializer and deserializer have been improved (after having seen actual
time on hardware).

The receiver was deemed too naive and protocol dependent and has subsequently
been dropped. Anyone using these pieces to build a receiver is encouraged to
build one to match their application.

Two test cases have been implemented to gain some confidence in the
encode(..)/decode(..) functions. This does not fully cover the decode(..) (see
comment on the test), but is a start.